backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
transformers
torch
sentencepiece
protobuf
hf_xet